# btcec: optimize square root using fieldVal